# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SmsLoginResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"walletExists": {
						"type": "boolean"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SmsLoginRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifySmsOTPResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"userData": {
								"type": "string"
							},
							"socialType": {
								"type": "string"
							},
							"walletAddress": {
								"type": "string"
							}
						},
						"required": [
							"userData",
							"socialType",
							"walletAddress"
						],
						"type": "object"
					},
					"isActive": {
						"type": "boolean"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifySmsOTPRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResendSmsOTPResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResendSmsOTPRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"app": {
				"type": "apiKey",
				"name": "X-App-ID",
				"in": "header",
				"description": "Application ID for API access"
			},
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT token for authentication"
			}
		}
	},
	"info": {
		"title": "crefy-connect-backend",
		"version": "1.0.0",
		"description": "Wallet authentication API with Express and TypeScript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jimleston Osoi"
		}
	},
	"paths": {
...
		"/auth/sms/login": {
			"post": {
				"operationId": "SmsLogin",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SmsLoginResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "OTP sent to phone number for verification",
											"isActive": false,
											"walletExists": false
										}
									}
								}
							}
						}
					}
				},
				"description": "Login or register with phone number\nSends OTP via SMS for verification",
				"tags": [
					"SMS Authentication"
				],
				"security": [
					{
						"app": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SmsLoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/sms/verify": {
			"post": {
				"operationId": "VerifySmsOTP",
				"responses": {
					"200": {
						"description": "Wallet verified successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifySmsOTPResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Wallet verified successfully",
											"data": {
												"walletAddress": "0x742E6B6D8B6C4e8D8e8D8e8D8e8D8e8D8e8D8e8D",
												"socialType": "sms",
												"userData": "{\"phoneNumber\":\"+1234567890\"}"
											},
											"isActive": true,
											"token": "jwt-token-here"
										}
									}
								}
							}
						}
					}
				},
				"description": "Verify SMS OTP and activate wallet",
				"tags": [
					"SMS Authentication"
				],
				"security": [
					{
						"app": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifySmsOTPRequest"
							}
						}
					}
				}
			}
		},
		"/auth/sms/resend-otp": {
			"post": {
				"operationId": "ResendSmsOTP",
				"responses": {
					"200": {
						"description": "OTP resent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResendSmsOTPResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "OTP resent to phone number for verification"
										}
									}
								}
							}
						}
					}
				},
				"description": "Resend OTP via SMS",
				"tags": [
					"SMS Authentication"
				],
				"security": [
					{
						"app": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResendSmsOTPRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}